backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
influxdb-client==1.47.0